array2ade4(made4)	R Documentation

Converts microarray input data into a data frame suitable for analysis in ADE4.

Description

Converts input data into a data.frame suitable for analysis in ADE4. This function is called by bga and other made4 function.

Usage

array2ade4(dataset, pos = FALSE, trans = TRUE)

Arguments

dataset 	 A matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

pos 	 Logical indicating whether to add an integer to dataset, to generate positive data.frame. Required for dudi.coa or dudi.nsc

trans 	 Logical indicating whether dataset should be transposed. Both between group analysis, and coinertia analysis expect that the cases (array samples) are contained in the rows and that the variables (genes) are in the columns. Thus the data.frame must be transposed prior to analysis

Details

bga and other functions in made4 call this function and it is generally not necessary to call array2ade4 this directly.

array2ade4 calls getdata, and will accept a matrix, data.frame, exprSet or marrayRaw format. It will also transpose data or add a integer to generate a positive data matrix.

If the input data contains missing values (NA), these must first be removed or imputed (see the R libraries impute() or pamr()).

Value

Returns a data.frame suitable for analysis by ade4 or made4 functions.

Author(s)

Aedin Culhane

References

See Also

getdata, getIntensityMatrix

Examples

library(affy)
data(geneData)
class(geneData)
dim(geneData)
dim(array2ade4(geneData))
class(array2ade4(geneData))

eset <- new("exprSet", exprs=geneData)
eset
class(eset)
class(array2ade4(eset))
dim(array2ade4(eset))

[Package Contents]