bet.coinertia(made4)	R Documentation

Between class coinertia analysis

Description

Between class coinertia analysis. cia of 2 datasets where covariance between groups or classes of cases, rather than individual cases are maximised.

Usage

bet.coinertia(df1, df2, fac1, fac2, cia.nf = 2, type = "nsc", ...)

Arguments

df1 	 First dataset.A matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

df2 	 Second dataset. A matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

fac1 	 A factor or vector which describes the classes in df1

fac2 	 A factor or vector which describes the classes in df2

cia.nf 	 Integer indicating the number of coinertia analysis axes to be saved. Default value is 2.

type 	 A character string, accepted options are type="nsc" or type="pca"

... 	 further arguments passed to or from other methods

Value

A list of class bet.cia of length 5

coin 	 An object of class 'coinertia', sub-class dudi. See coinertia

coa1, pca1 	 An object of class 'nsc' or 'pca', with sub-class 'dudi'. See dudi, dudi.pca or dudi.nsc

bet1 	 An object of class 'bga', with sub-class 'dudi'. See dudi, bga or between

bet2 	 An object of class 'bga', with sub-class 'dudi'. See dudi, bga or between.

Note

This is very computational intensive. The authors of ade4 are currently re-writing the code for coinertia analysis, so that it should substantially improve the computational requirements (May 2004).

Author(s)

Aedin Culhane

References

Culhane AC, et al., 2003 Cross platform comparison and visualisation of gene expression data using co-inertia analysis. BMC Bioinformatics. 4:59

See Also

See Also as coinertia, cia.

Examples

### NEED TO DO 
if (require(ade4, quiet = TRUE)) {}

[Package Contents]