between.graph(made4)	R Documentation

Plot 1D graph of results from between group analysis

Description

Plots a 1D graph, of results of between group analysis similar to that in Culhane et al., 2002.

Usage

between.graph(x, ax=1, cols=NULL, hor=TRUE, ...)

Arguments

x 	 Object of the class bga resulting from a bga analysis.

ax 	 Numeric. The column number of principal component ($ls and $li) to be used. Default is 1. This is the first component of the analysis.

cols 	 Vector of colours. By default colours are obtained using getcol

hor 	 Logical, indicating whether the graph should be plotted horizontally or vertically. The default is a horizontal plot.

... 	 further arguments passed to or from other methods

Details

This will produce a figure similar to Figure 1 in the paper by Culhane et al., 2002.

between.graph requires both samples and centroid co-ordinates ($ls, $li) which are passed to it via an object of class bga. If cases are to be coloured by class, it also requires a $fac factor which is also passed to it via an object of class bga.

To plot a 1D graph from other multivariate analysis such as PCA (dudi.pca), COA (dudi.coa), or coinertia analysis. Please use graph1D.

Value

Note

Author(s)

Aedin Culhane

References

Culhane AC, et al., 2002 Between-group analysis of microarray data. Bioinformatics. 18(12):1600-8.

See Also

graph1D

Examples

data(khan)
if (require(ade4, quiet = TRUE)) {
khan.bga<-bga(khan$train, khan$train.classes)
}

between.graph(khan.bga)
between.graph(khan.bga, ax=2, lwd=3, hor=FALSE, cex=0.5, col=c("green", 
"blue", "red", "yellow"))

[Package Contents]