bga.jackknife(made4)	R Documentation

Jackknife between group analysis

Description

Performs one-leave-out jackknife analysis of a between group analysis as described by Culhane et al., 20002

Usage

bga.jackknife(data, classvec, ...)

Arguments

data 	 Input dataset. A matrix, data.frame If the input is gene expression data in a matrix or data.frame. The columns contain the cases (array samples) which will be jackknifed.

classvec 	 A factor or vector which describes the classes in the training dataset

... 	 further arguments passed to or from other methods

Details

Performs a one-leave-out cross validation of between group analysis bga. Input is a training dataset. This can take 5-10 minutes to compute on standard data gene expression matrix.

In jackknife one leave out analysis, one case (column) is removed. The remaining dataset is subjected to bga. Then the class of the case that was removed is predicted using suppl. This analysis is repeated until all samples have been removed and predicted.

Value

A list containing

results 	 The projected co-ordinates of each sample

summary 	 A summary of number and percentage of correctly assigned samples

Author(s)

Aedin Culhane

References

Culhane et al., 2002 Between-group analysis of microarray data. Bioinformatics. 18(12):1600-8.

See Also

See Also bga, bga.suppl, suppl,bga, between, plot.bga

Examples

data(khan)
# NOTE using a very reduced dataset (first 5 genes) to speed up results
# hence expect poor prediction accuracy
dim(khan$train)
print("using only small subset of data")
if (require(ade4, quiet = TRUE)) {
bga.jackknife(khan$train[1:5,], khan$train.classes) }

[Package made4 version 0.5 Index]