bga.suppl(made4)	R Documentation

Between group analysis with supplementary data projection

Description

bga.suppl performs a bga between group analysis with projection of supplementary points using suppl

Usage

bga.suppl(dataset, supdata, classvec, supvec = NULL, suponly = FALSE,...)

Arguments

dataset 	 Training dataset. A matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

supdata 	 Test or blind dataset. A matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively. The test dataset supdata and the training dataset dataset must contain the same number of variables (genes).

classvec 	 A factor or vector which describes the classes in the training data dataset

supvec 	 A factor or vector which describes the classes in the test dataset supdata

suponly 	 Logical indicating whether the returned output should contain the test class assignment results only. The default value is FALSE, that is the training coordinates, test coordinates and class assignments will all be returned.

... 	 further arguments passed to or from other methods

Details

bga.suppl calls bga to perform between group analysis (bga) on the training dataset, then it calls suppl to project the test dataset onto the bga axes. It returns the coordinates and class assignment of the cases (microarray samples) in the test dataset as described by Culhane et al., 2002.

The test dataset must contain the same number of variables (genes) as the training dataset.

The input format of both the training dataset and test dataset are verified using array2ade4. Use plot.bga to plot results from bga.

Value

If suponly is FALSE (the default option) bga.suppl returns a list of length 4 containing the results of the bga of the training dataset and the results of the projection of the test dataset onto the bga axes-

ord 	 Results of initial ordination. A list of class "dudi" (see dudi).

bet 	 Results of between group analysis. A list of class "dudi" (see dudi),"between" (see between),and "dudi.bga"(see bga)

fac 	 The input classvec, the factor or vector which described the classes in the input dataset

suppl 	 An object returned by suppl

If suponly is TRUE only the results from suppl will be returned.

Author(s)

Aedin Culhane

References

Culhane AC, et al., 2002 Between-group analysis of microarray data. Bioinformatics. 18(12):1600-8.

See Also

See Also bga, suppl, between, plot.bga, bga.jackknife

Examples

data(khan)
#khan.bga<-bga(khan$train, khan$train.classes)
if (require(ade4, quiet = TRUE)) {
khan.bga<-bga.suppl(khan$train, supdata=khan$test, 
classvec=khan$train.classes, supvec=khan$test.classes)

khan.bga
plot.bga(khan.bga, genelabels=khan$annotation$Symbol)
khan.bga$suppl
}

[Package made4 version 0.5 Index]