commonMap(made4)	R Documentation

Highlight common points between two 1D plots

Description

CommonMap draws two 1D plots, and links the common points between the two.

Usage

commonMap(x, y, hor=TRUE, cex=1.5, scaled=TRUE, ...)

Arguments

x 	 The coordinates of the first axis

y 	 The coordinates of the second axis

hor 	 Logical, whether a horizontal line should be drawn on plot. Default is TRUE.

cex 	 Numeric. The amount by which plotting text and symbols should be scaled relative to the default

scaled 	 Logical, whether the data in x and y are scaled. Scaling is useful for visualising small or large data values. Set to FALSE if actually or true values should be visualised. The default is TRUE.

... 	 further arguments passed to or from other method

Details

Useful for mapping the genes in common from coinertia analysis This graphs a 1D graph, x and y are the coordinates from two different analyses but the rows of each vectors correspond (ie common genes)

Value

Note

This is useful for examining common points in axes from coinertia analysis, or comparing results from two different analysis.

Author(s)

Ailis Fagan and Aedin Culhane

References

See Also

See also between.graph, graph1D

Examples

a<-rnorm(20)
b<-rnorm(20)
par(mfrow=c(2,2))
commonMap(a,b)
commonMap(a,b,hor=FALSE, col="red", pch=19)
commonMap(a,b,col="blue", cex=2, pch=19)

# If the vectors contain different variables, the rows should define the variables that correspond
a[15:20]<-NA
b[10:15]<-NA
cbind(a,b)
commonMap(a,b, col="dark green", pch=18)

[Package Contents]