dudi.rwcoa(made4)	R Documentation

Row weighted Correspondence Analysis

Description

dudi.rwcoa Row weighted COA, calls forrwcoa to perform row weighted correspondence analysis.

Usage

dudi.rwcoa(df, rowweights = rep(1/nrow(df),nrow(df)), ...)

Arguments

df 	 a data.frame containing positive or null values. It should not contain missing (NA) values.

rowweights 	 a vector of row weights (by default, uniform row weights)

... 	 further arguments passed to or from other methods )

Details

Performs row weighted COA. Calls forrwcoa to calculates weights.

Value

Returns a list of class 'coa', 'rwcoa', and 'dudi' (see dudi)

Note

In the paper by Culhane et al., 2002, coinertia analysis was performed with two COAs, a standard COA and a row weighted COA dudi.rwcoa, on the two gene expression datasets. However it is now recommended to perform two non-symmetric COA, instead of two COA. This avoids having to force the row weights from one analysis on the second. To perform non-symmetric correspondence coinertia analysis, use nsc.coinertia.

Author(s)

Aedin Culhane, A.B. Dufour

References

Culhane AC, et al., 2003 Cross platform comparison and visualisation of gene expression data using co-inertia analysis. BMC Bioinformatics. 4:59

See Also

See Also as dudi,dudi.coa,dudi.pca nsc.coinertia

Examples

if (require(ade4, quiet = TRUE)) {
data(atlas)
coa1=dudi.coa(atlas$birds,scannf=FALSE)
coa2=dudi.coa(forrwcoa(atlas$birds),scannf=FALSE)
coa3=dudi.rwcoa(atlas$birds, rowweights=coa1$lw, scannf=FALSE)
}

[Package Contents]