genes(made4)	R Documentation

Returns an index of the top N variables at the ends of the x and y axes

Description

Returns an index of the top n maximum and minimum values in 2 columns of a matrix or data.frame.

Usage

genes(dudivar,n=5, axes1=1, axes2=2)

Arguments

dudivar 	 a data frame or matrix with at least two columns, containing the x,y coordinates.

axes1 	 the column number for the x-axis~

axes2 	 the column number for the y-axis~

n

{n is the number of genes to be extracted}

Details

Genes returns an index of the "top" variables at the ends of the x and y axes. Genes is not normally called by the user but is called by plotgenes()

If you wish to return a table or list of the top genes at the end of an axis, use the function topgenes.

Thanks to Miss Jess Mar for helping on this function.

Value

Author(s)

Aedin Culhane

See Also

See Also as plotgenes

Examples

[Package Contents]