genes1d(made4)	R Documentation

genes1d: list top N variables on one axis. Lists the maximum and minimum values in a column of a data.frame

Description

Lists the top N variables from the positive and negative ends of an axis. Returns a list of N variables that have the max and min. Given a $co or $li file it will return that variables at the ends of the axis.

Usage

genes1d(array, n = 5, axis = 1, listgenes = FALSE)

Arguments

array 	 A matrix or data.frame. Typically a $co or $li from bga, cia or dudi.

n 	 An integer indicating the number of variables to be returned. Default is 5.

axis 	 An integer indicating the column of x. Default is 1 (first axis, of $co or $li file)

listgenes 	 A logical, If TRUE a list containing the variables at each ends of the axes are returned

... 	 further arguments passed to or from other methods

Details

Genes1d is similar to link[made4:genes]{genes}, but returns an index of genes at the ends of one axes.

Value

Returns a vector or list of vectors.

Author(s)

Aedin Culhane

References

See Also

See Also as genes

Examples

data(khan)
if (require(ade4, quiet = TRUE)) {
khan.coa<-dudi.coa(khan$train[1:100,1:10], scannf=FALSE, nf=2)
}
ind<-topgenes(khan.coa$li, ends="pos")
ind.ID<-topgenes(khan.coa$li, ends="pos", labels=khan$gene.labels.imagesID)
ind.symbol<-topgenes(khan.coa$li, ends="pos", labels=khan$annotation$Symbol)
Top10.pos<- cbind("Gene Symbol"=ind.symbol, 
"Clone ID"=ind.ID, "Coordinates"=khan.coa$li[ind,], row.names=c(1:length(ind)))
Top10.pos

[Package Contents]