getcol(made4)	R Documentation

Specialised colour palette with set of 21 maximally contrasting colours

Description

Special colour palette developed to maximise the contrast between colours. Colours were selected for visualising groups of points on xy or xyz plots on a white background. Because of this, there are few pastel colours are in this palette. getcol contains 2 palettes of 9 and 21 colours.

Usage

getcol(nc = c(1:3), palette = "colours2", test = FALSE)

Arguments

nc 	 Numeric. Integer or vector in range 1 to 21. This selects colours from palette

palette 	 A character to select either palette "colours1" or "colours2". colours1 contains 9 colours, colours2 contains 21 colours

test 	 A logical, if TRUE a plot will be drawn to display the palettes colours1, colours2 and any selected colours.

Details

Colours1 contains the 9 colours,"red","blue" ,"green","cyan","magenta","yellow", "grey","black","brown". Colours2 contains 21 colours. These were selected so as to maximise the contrast between groups.

For other colour palettes in R, see colors, palette, rainbow, heat.colors, terrain.colors, topo.colors or cm.colors.

Also see the library RColorBrewer

Value

A vector containing a list of colours.

Author(s)

Aedin Culhane

See Also

See also colors, palette, rainbow, heat.colors, terrain.colors, topo.colors or cm.colors, RColorBrewer

Examples

getcol(3)
getcol(c(1:7))
getcol(10, test=TRUE)
getcol(c(1:5, 7, 15, 16), palette="colours2",test=TRUE)

[Package Contents]