getdata(made4)	R Documentation

Check dataset

Description

This function is called by array2ade4 which converts input data into a data.frame suitable for analysis in ADE4. This function is not normally called by the user

Usage

getdata(arraydata)

Arguments

arraydata 	 a matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

Details

bga and other functions in made4 call this function and it is generally not necessary to call this directly.

array2ade4 calls getdata, and will accept a matrix,data.frame, exprSet or marrayRaw format. It will also transpose data or add a integer to generate a positive data matrix. If the input data contains NA, these must first be removed or imputed (see the R libraries impute() or pamr()). getdata is very similar to getIntensityMatrix in vsn.

Value

Returns a data.frame suitable for analysis by ade4 or made4 functions.

Note

This is not normally called by a user. Internal function.

Author(s)

Aedin Culhane

See Also

array2ade4, getIntensityMatrix

Examples

[Package Contents]