graph1D(made4)	R Documentation

Plot 1D graph of axis from multivariate analysis

Description

Draw 1D plot of an axis from multivariate analysis. Useful for visualising an individual axis from analyses such as PCA dudi.pca or COA dudi.coa. It accepts a factor so that groups of points can be coloured. It can also be used for graphing genes, and will only label n genes at the ends of the axis.

Usage

graph1D(dfx,  fac=NULL,ax = 1, hor=FALSE, s.nam=row.names(dfx), n=NULL, 
       scaled=TRUE, col="red", width=NULL, ...)

Arguments

dfx 	 vector, matrix, or data.frame, which contains a column with axis coordinates

ax 	 Numeric, indicating column of matrix, or data.frame to be plotted. The default is 1.

fac 	 Factor, indicating sub-groupings or classes in dfx or dfx[,ax]

hor 	 Logical, indicating whether the graph should be drawn horizontal or vertically. The default is vertically.

s.nam 	 Vector. labels of dfx, The default is row.names(dfx)

n 	 Numeric. Whether all rows should be plotted, n=10 would label only the 10 variables at the end of the axis. By default all variables (row of dfx) are labelled

scaled

col 	 A character or vector indicating the colour(s) for points or groups of points. If points are to be coloured according to a factor, length(col) should equal length(levels(fac))

width 	 A vector of length 2, which is the width (of a vertical plot) or height (of a horizontal plot). This can be increased if variable labels are unreadable. The default is c(-2,1)

... 	 further arguments passed to or from other methods

Details

Value

Note

Author(s)

Aedin Culhane

References

See Also

between.graph

Examples

a<-rnorm(25)
graph1D(a, s.nam=letters[1:25])
graph1D(a, s.nam=letters[1:25], col="blue", pch=19,  n=3)
data(khan)
if (require(ade4, quiet = TRUE)) {
khan.coa<-dudi.coa(khan$train, scan=FALSE, nf=2)
}
graph1D(khan.coa$co, ax=1)

[Package made4 version 0.5 Index]