heatplot(made4)	R Documentation

Draws heatmap with dendrograms.

Description

heatplot calls heatmap using a red-green colour scheme by default. It also draws dendrograms of the cases and variables using correlation similarity metric and average linkage clustering as described by Eisen. heatplot is useful for a quick overview or exploratory analysis of data

Usage

heatplot(dataset, dend = TRUE, lowcol = "green", 
                highcol = "red", Colv=NULL, Rowv=NULL, ...)

Arguments

dataset 	 a matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

dend 	 Logical, indicating whether dendrograms should be drawn. Default is TRUE. If FALSE both Colv and Rowv are set to NA.

Colv, Rowv 	 Vector or object of class dendrogram used to reorder the columns, or rows. If no ordering is required, set Colv or Rowv = NA. The default is NULL

lowcol, highcol 	 Character indicating colours to be used for down and upregulated genes when drawing heatmap. Default is lowcol="green", and highcol="red".

... 	 further arguments passed to or from other methods

Details

The hierarchical plot is produced using average linkage cluster analysis with a correlation metric distance. heatplot calls heatmap and dendrogram.

Value

Note

Because Eisen et al., 1998 use green-red colours for the heatmap heatplot uses these by default however a blue-red or yellow-blue are easily obtained by changing lowcol and highcol

Author(s)

Aedin Culhane

References

Eisen MB, Spellman PT, Brown PO and Botstein D. (1998). Cluster Analysis and Display of Genome-Wide Expression Patterns. Proc Natl Acad Sci USA 95, 14863-8.

See Also

See also as hclust, heatmap and dendrogram

Examples

data(khan)

heatplot(khan$train[1:30,], lowcol="blue", highcol="red")
heatplot(khan$train[1:26,], lowcol="blue", highcol="red", 
         labRow = c(64:1), labCol=LETTERS[1:26])

if (require(ade4, quiet = TRUE)) {
 # To speed up analysis, only a subset of the genes are analysed
khan.sub<-array2ade4(khan$train[1:1000,])  
khan.coa<-dudi.coa(khan.sub, scan=FALSE, nf=6)
}

# Provides a view of the components of the Correspondence analysis 
heatplot(khan.coa$li, dend=FALSE)   

# transposed so that it is easier to view. Can see that the difference between tissues 
# and cell line samples are defined in the first axis.

heatplot(t(khan.coa$li), dend=FALSE, lowcol="blue") 

[Package Contents]