ord(made4)	R Documentation

Ordination

Description

Run principal component analysis, correspondence analysis or non-symmetric correspondence analysis on gene expression data

Usage

ord(dataset, type="coa", classvec=NULL, ...)
plot.ord(x, axes1=1, axes2=2, arraycol=NULL, genecol="gray25", nlab=10, genelabels= NULL, classvec=NULL, ...)

Arguments

dataset 	 Training dataset. A matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

classvec 	 A factor or vector which describes the classes in the training dataset

type 	 Character, "coa", "pca" or "nsc" indicating which data transformation is required. The default value is type="coa"

x 	 An object of class ord. The output from ord. It contains the projection coordinates from ord, the $co or $li coordinates to be plotted

arraycol, genecol 	 Character, colour of points on plot. If arraycol is NULL, arraycol will obtain a set of contrasting colours using getcol, for each classes of cases (microarray samples) on the array (case) plot. genecol is the colour of the points for each variable (genes) on gene plot

nlab 	 Numeric. An integer indicating the number of variables (genes) at the end of axes to be labelled, on the gene plot.

axes1 	 Integer, the column number for the x-axis. The default is 1.

axes2 	 Integer, the column number for the y-axis, The default is 2.

genelabels 	 A vector of variables labels, if genelabels=NULL the row.names of input matrix dataset will be used

... 	 further arguments passed to or from other methods

Details

ord calls either dudi.pca, dudi.coa or dudi.nsc on the input dataset. The input format of the dataset is verified using array2ade4.

If the user defines microarray sample groupings, these are colours on plots produced by plot.ord.

Plotting and visualising bga results:

2D plots: s.var and s.groups to draw an xy plot of cases ($ls). s.var and s.groups are modifications of the ADE4 graphing functions s.label and s.class. plotgenes, is used to draw an xy plot of the variables (genes).

3D plots: 3D graphs can be generated using do3D and html3D. html3D produces a web page in which a 3D plot can be interactively rotated, zoomed, and in which classes or groups of cases can be easily highlighted.

1D plots, show one axis only: 1D graphs can be plotted using graph1D. graph1D can be used to plot either cases (microarrays) or variables (genes) and only requires a vector of coordinates ($li, $co)

Analysis of the distribution of variance among axes:

The number of axes or principal components from a ord will equal nrow the number of rows, or the ncol, number of columns of the dataset (whichever is less).

The distribution of variance among axes is described in the eigenvalues ($eig) of the ord analysis. These can be visualised using a scree plot, using scatterutil.eigen as it done in plot.ord. It is also useful to visualise the principal components from a using a ord or principal components analysis dudi.pca, or correspondence analysis dudi.coa using a heatmap. In MADE4 the function heatplot will plot a heatmap with nicer default colours.

Extracting list of top variables (genes):

Use topgenes to get list of variables or cases at the ends of axes. It will return a list of the top n variables (by default n=5) at the positive, negative or both ends of an axes. sumstats can be used to return the angle (slope) and distance from the origin of a list of coordinates.

Value

A list with a class ord containing:

ord 	 Results of initial ordination. A list of class "dudi" (see dudi )

fac 	 The input classvec, the factor or vector which described the classes in the input dataset. Can be null.

Author(s)

Aedin Culhane

References

See Also

See Also dudi.pca, dudi.coa or dudi.nsc, bga,

Examples

data(khan)

if (require(ade4, quiet = TRUE)) {
  khan.coa<-ord(khan$train, classvec=khan$train.classes, type="coa")  
  }

khan.coa
plot(khan.coa, genelabels=khan$annotation$Symbol)

# Provide a view of the first 5 principal components (axes) of the correspondence analysis
heatplot(khan.coa$ord$li[,1:5], dend=FALSE,lowcol="blue")

[Package made4 version 0.52 Index]