overview(made4)	R Documentation

Draw boxplot histogram and hierarchical tree of gene expression data

Description

Very simple wrapper function that draws a boxplot, histogram and hierarchical tree of expression data

Usage

overview(dataset, label=NULL, title="", ...)

Arguments

dataset 	 a matrix, data.frame, exprSet or marrayRaw. If the input is gene expression data in a matrix or data.frame. The rows and columns are expected to contain the variables (genes) and cases (array samples) respectively.

label 	 Vector, labels to be placed on samples in plots. Default is rownames(dataset)

title 	 Character, label to be placed on plots. Default is NULL

... 	 further arguments passed to or from other methods

Details

The hierarchical plot is produced using average linkage cluster analysis with Eisen's correlation metric.

Value

Note

Author(s)

Aedin Culhane

References

See Also

See also as boxplot, hclust, hist

Examples

  data(khan)

  subkhan<-log2(khan$train[1:1000,])
  overview(subkhan, title="Subset of Khan Training data")

[Package Contents]