s.var(made4)	R Documentation

Draw xy plots of results of multivariate analysis

Description

A number of functions for graphically displaying results from between group, coinertia, or other multivariate analysis of gene expression data. s.var, s.groups and s.match.col are derived from ADE4 graphics modules, s.label, s.class and s.match respectively.

Usage

s.var(dfxy, xax = 1, yax = 2, label = row.names(dfxy), clabel = 1, pch = 20, 
      cpoint = if (clabel == 0) 1 else 0, neig = NULL, cneig = 2, xlim = NULL, 
      ylim = NULL, grid = TRUE, addaxes = TRUE, cgrid = 1, include.origin = TRUE, 
      origin = c(0, 0), sub = "", csub = 1.25, possub = "bottomleft", pixmap = NULL, 
      contour = NULL, area = NULL, add.plot = FALSE, colpoints = rep(1, nrow(dfxy)), 
      boxes = TRUE, ...)
s.groups(dfxy, fac, wt = rep(1, length(fac)), xax = 1, yax = 2, cstar = 1, 
      cellipse = 1.5, axesell = TRUE, label = levels(fac), slabel = row.names(dfxy), 
      clabel = 1, cpoint = 1, pch = 20, col = rep(1, length(levels(fac))), 
      xlim = NULL, ylim = NULL, grid = TRUE, addaxes = TRUE, origin = c(0, 0), 
      include.origin = TRUE, sub = "", csub = 1, possub = "bottomleft", 
      cgrid = 1, pixmap = NULL, contour = NULL, area = NULL, add.plot = FALSE, ...)
s.match.col(df1xy, df2xy, xax = 1, yax = 2, pch = 20, cpoint = 1, 
    label = row.names(df1xy), clabel = 1, edge = TRUE, xlim = NULL, 
    ylim = NULL, grid = TRUE, addaxes = TRUE, cgrid = 1, include.origin = TRUE, 
    origin = c(0, 0), sub = "", csub = 1.25, possub = "bottomleft", 
    pixmap = NULL, contour = NULL, area = NULL, add.plot = FALSE, col=rep(1, nrow(df1xy)), 
    fac=NULL,  ...) 

Arguments

dfxy 	 coordinates, normally the $ls, $co or $li coordinates from PCA, COA or BGA with at least two columns, containing the x,y coordinates to be plotted

df1xy, df2xy 	 coordinates from cia, normally the $mX, $mY coordinates with at least two columns, containing the x,y coordinates to be plotted

colpoints 	 Character, colour of arrays (cases),genes or points

xax 	 the column number for the x-axis

yax 	 the column number for the y-axis

fac 	 Factor. A factor, classvec partitioning the rows of the data frame in classes

boxes 	 A logical, indicating whether a box should be drawn around each variable label. The default is TRUE

add.plot 	 logical if TRUE plot draw in current graphics window

origin, sub, wt, cstar, cellipse, axesell, col, edge 	 Graphing parameters. See s.label or s.class for details

clabel,pch, cpoint, neig, cneig, xlim, ylim, grid, addaxes, cgrid, include.origin, csub, possub, pixmap, contour, area 	 Graphing parameters, see s.match for details

... 	 further arguments passed to or from other method

Details

These functions are modified versions of s.label, s.class and s.match graph modules from ADE4. Please see s.label or s.class and s.match for father details.

These functions s.var, s.groups and s.match.col are called by plot.bga and plot.cia.

Author(s)

Aedin Culhane

See Also

See also s.class, s.label, s.match

Examples

data(khan)
if (require(ade4, quiet = TRUE)) {
khan.bga<-bga(khan$train, khan$train.classes)
attach(khan.bga)
par(mfrow=c(2,2))
s.var(bet$ls, col=as.numeric(khan$train.classes), clabel=0.8)
s.groups(bet$ls, khan$train.classes, add.plot=TRUE, 
col=c(1:length(levels(khan$train.classes))) )

}

[Package made4 version 0.52 Index]