suppl(made4)	R Documentation

Projection of supplementary data onto axes from a between group analysis

Description

Projection and class prediction of supplementary points onto axes from a between group analysis, bga.

Usage

suppl(dudi.bga, supdata, supvec = NULL, assign=TRUE, ...)

Arguments

dudi.bga 	 An object returned by bga

supdata 	 Test or blind dataset. Accepted formats are a matrix, data.frame, exprSet or marrayRaw

supvec 	 A factor or vector which describes the classes in the training dataset

assign 	 A logical indicating whether class assignment should be calculated using the method described by Culhane et al., 2002. The default value is TRUE

... 	 further arguments passed to or from other methods

Details

After performing a between group analysis on a training dataset using bga, a test dataset can be then projected onto bga axes using suppl.

suppl returns the projected coordinates and assignment of each test case (array).

The test dataset must contain the same number of variables (genes) as the training dataset. The input format of both the training dataset and test dataset are verified using array2ade4. Use plot.bga to plot results from bga.

Value

A list containing:

suppl 	 An object returned by suppl

Author(s)

Aedin Culhane

References

Culhane AC, et al., 2002 Between-group analysis of microarray data. Bioinformatics. 18(12):1600-8.

See Also

See Also bga, between, plot.bga, bga.jackknife

Examples

data(khan)
#khan.bga<-bga(khan$train, khan$train.classes)
if (require(ade4, quiet = TRUE)) {
khan.bga<-bga.suppl(khan$train, supdata=khan$test, classvec=khan$train.classes, 
          supvec=khan$test.classes)

khan.bga
plot.bga(khan.bga, genelabels=khan$annotation$Symbol)
khan.bga$suppl
}

[Package Contents]