topgenes(made4)	R Documentation

Topgenes returns a list of variables at the ends (positive negative or both) of an axis

Description

topgenes will return a list of the top N variables from the positive, negative or both ends of an axis. That is, it returns a list of variables that have the maximum and/or minimum values in a vector.

Usage

topgenes(x, n = 10, axis = 1, labels = row.names(x), ends = "both", ...)

Arguments

x 	 A vector, matrix or data.frame. Typically a data frame $co or $li from dudi or $ls, $li, $co from bga.

n 	 An integer indicating the number of variables to be returned. Default is 5.

axis 	 An integer indicating the column of x. Default is 1 (first axis, of $co or $li file)

labels 	 A vector of row names, for x[,axis]. Default values is row.names(x)

ends 	 A string, "both", "neg", "pos", indicating whether variable label should be return from both, the negative or the positive end of an axis. The default is both.

... 	 further arguments passed to or from other methods

Details

topgenes calls genes1d. genes1d is similar to genes, but returns an index of genes at the ends of one axis not two axes. Given a $co or $li file it will return that variables at the ends of the axis.

Value

Returns a vector or list of vectors.

Note

Author(s)

AedinCulhane

References

See Also

See Also as genes

Examples

# Simple example
a<-rnorm(50)
order(a)
topgenes(a, labels=c(1:length(a)), ends="neg")

# Applied example
data(khan)
if (require(ade4, quiet = TRUE)) {
khan.coa<-dudi.coa(khan$train[1:100,1:10], scannf=FALSE, nf=2)
}
ind<-topgenes(khan.coa$li, ends="pos")
ind.ID<-topgenes(khan.coa$li, ends="pos", labels=khan$gene.labels.imagesID)
ind.symbol<-topgenes(khan.coa$li, ends="pos", labels=khan$annotation$Symbol)
Top10.pos<- cbind("Gene Symbol"=ind.symbol, 
      "Clone ID"=ind.ID, "Coordinates"=khan.coa$li[ind,], row.names=c(1:length(ind)))
Top10.pos

[Package Contents]